# frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Vertiche LegalOps API",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/documentos": {
            "get": {
                "tags": [
                    "documentos"
                ],
                "summary": "List Documentos",
                "operationId": "list_documentos",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "title": "Page"
                        }
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 25,
                            "title": "Page Size"
                        }
                    },
                    {
                        "name": "estado_ocr",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Estado Ocr"
                        }
                    },
                    {
                        "name": "requiere_revision",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "boolean"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Requiere Revision"
                        }
                    },
                    {
                        "name": "tienda_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Tienda Id"
                        }
                    },
                    {
                        "name": "tramite_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Tramite Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedResponse_Documento_"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/documentos/{id}": {
            "get": {
                "tags": [
                    "documentos"
                ],
                "summary": "Get Documento",
                "operationId": "get_documento",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Documento"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "documentos"
                ],
                "summary": "Update Documento",
                "operationId": "update_documento",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DocumentoUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Documento"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "documentos"
                ],
                "summary": "Delete Documento",
                "operationId": "delete_documento",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/documentos/{id}/url": {
            "get": {
                "tags": [
                    "documentos"
                ],
                "summary": "Get Documento Url",
                "operationId": "get_documento_url",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    },
                    {
                        "name": "download",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "boolean",
                            "title": "Download"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentoUrl"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/documentos/{id}/rename": {
            "post": {
                "tags": [
                    "documentos"
                ],
                "summary": "Rename Documento",
                "operationId": "rename_documento",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DocumentoRename"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Documento"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/documentos/{id}/ocr-review": {
            "post": {
                "tags": [
                    "documentos"
                ],
                "summary": "Accept Ocr Review",
                "operationId": "accept_ocr_review",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DocumentoOcrReview"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Documento"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/tiendas/{id}/documentos": {
            "get": {
                "tags": [
                    "tiendas"
                ],
                "summary": "Get Documentos For Tienda",
                "operationId": "get_documentos_for_tienda",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "title": "Page"
                        }
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 25,
                            "title": "Page Size"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedResponse_Documento_"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "tiendas"
                ],
                "summary": "Upload Documento For Tienda",
                "operationId": "upload_documento_for_tienda",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_upload_documento_for_tienda_api_tiendas__id__documentos_post"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Documento"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Body_upload_documento_for_tienda_api_tiendas__id__documentos_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "contentMediaType": "application/octet-stream",
                        "title": "File"
                    },
                    "file_name": {
                        "type": "string",
                        "title": "File Name"
                    },
                    "tramite_ids": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Tramite Ids"
                    }
                },
                "type": "object",
                "title": "Body_upload_documento_for_tienda_api_tiendas__id__documentos_post"
            },
            "CampoExtraidoDocumento": {
                "properties": {
                    "value": {
                        "type": "string",
                        "title": "Value"
                    },
                    "confidence": {
                        "type": "number",
                        "title": "Confidence"
                    }
                },
                "type": "object",
                "required": [
                    "value",
                    "confidence"
                ],
                "title": "CampoExtraidoDocumento"
            },
...
            "Documento": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "tramite_ids": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Tramite Ids"
                    },
                    "tramite_nombres": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tramite Nombres"
                    },
                    "nombre_archivo": {
                        "type": "string",
                        "title": "Nombre Archivo"
                    },
                    "ruta_archivo": {
                        "type": "string",
                        "title": "Ruta Archivo"
                    },
                    "estado_ocr": {
                        "$ref": "#/components/schemas/EstadoOCR"
                    },
                    "datos_extraidos": {
                        "anyOf": [
                            {
                                "additionalProperties": {
                                    "$ref": "#/components/schemas/CampoExtraidoDocumento"
                                },
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Datos Extraidos"
                    },
                    "requiere_revision_manual": {
                        "type": "boolean",
                        "title": "Requiere Revision Manual"
                    },
                    "cargado_por": {
                        "type": "string",
                        "title": "Cargado Por"
                    },
                    "cargado_por_nombre": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Cargado Por Nombre"
                    },
                    "cargado_en": {
                        "type": "string",
                        "title": "Cargado En"
                    },
                    "tienda_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tienda Id"
                    },
                    "tienda_nombre": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tienda Nombre"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "tramite_ids",
                    "nombre_archivo",
                    "ruta_archivo",
                    "estado_ocr",
                    "requiere_revision_manual",
                    "cargado_por",
                    "cargado_en"
                ],
                "title": "Documento"
            },
            "DocumentoOcrReview": {
                "properties": {
                    "datos_extraidos": {
                        "additionalProperties": {
                            "type": "string"
                        },
                        "type": "object",
                        "title": "Datos Extraidos"
                    }
                },
                "type": "object",
                "required": [
                    "datos_extraidos"
                ],
                "title": "DocumentoOcrReview"
            },
            "DocumentoRename": {
                "properties": {
                    "nombre_archivo": {
                        "type": "string",
                        "title": "Nombre Archivo"
                    }
                },
                "type": "object",
                "required": [
                    "nombre_archivo"
                ],
                "title": "DocumentoRename"
            },
            "DocumentoUpdate": {
                "properties": {
                    "tramite_ids": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tramite Ids"
                    }
                },
                "type": "object",
                "title": "DocumentoUpdate"
            },
            "DocumentoUrl": {
                "properties": {
                    "url": {
                        "type": "string",
                        "title": "Url"
                    }
                },
                "type": "object",
                "required": [
                    "url"
                ],
                "title": "DocumentoUrl"
            },
...
            "EstadoOCR": {
                "type": "string",
                "enum": [
                    "procesando",
                    "completado",
                    "baja_confianza",
                    "error"
                ],
                "title": "EstadoOCR"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "PaginatedResponse_Documento_": {
                "properties": {
                    "data": {
                        "items": {
                            "$ref": "#/components/schemas/Documento"
                        },
                        "type": "array",
                        "title": "Data"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    },
                    "page": {
                        "type": "integer",
                        "title": "Page"
                    },
                    "page_size": {
                        "type": "integer",
                        "title": "Page Size"
                    },
                    "total_pages": {
                        "type": "integer",
                        "title": "Total Pages"
                    }
                },
                "type": "object",
                "required": [
                    "data",
                    "total",
                    "page",
                    "page_size",
                    "total_pages"
                ],
                "title": "PaginatedResponse[Documento]"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/api/auth/login"
                    }
                }
            }
        }
    }
}